MCKIN-2215 fix result to be float, to be sure this is getting calculated...

+++ admin/controller.py
@@ -445,7 +445,7 @@ def get_organizations_users_completion(client_id, course_id, users_enrolled):
     users_completed = organization_api.get_grade_complete_count(client_id, courses=course_id).users_grade_complete_count
     percent_completed = '0%'
     if users_enrolled > 0:
-        percent_completed = "{}%".format(int(int(users_completed) / int(users_enrolled) * 100))
+        percent_completed = "{}%".format(int((float(users_completed) / float(users_enrolled)) * 100))
     return users_completed, percent_completed
